auth_service/requirements.txt
fastapi==0.112.0
python-jose==3.3.0
pytest-mock==3.14.0